backend/requirements.txt
asgiref==3.8.1
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
